kt/src/commonTest/kotlin/com/github/ephemient/aoc2022/Day5Test.kt
package com.github.ephemient.aoc2022

import kotlin.test.Test
import kotlin.test.assertEquals

class Day5Test {
    @Test
    fun part1() {
        assertEquals("CMZ", Day5(SAMPLE_INPUT).part1())
    }

    @Test
    fun part2() {
        assertEquals("MCD", Day5(SAMPLE_INPUT).part2())
    }

    companion object {
        private val SAMPLE_INPUT = listOf(
            "    [D]    ",
            "[N] [C]    ",
            "[Z] [M] [P]",
            " 1   2   3 ",
            "",
            "move 1 from 2 to 1",
            "move 3 from 1 to 3",
            "move 2 from 2 to 1",
            "move 1 from 1 to 2"
        )
    }
}
